+++ base_file/.gitignore
@@ -1,0 +1,1 @@
+Cargo.lock
