+++ base_file/Cargo.toml
@@ -1,0 +1,13 @@
+[package]
+name = "base_file"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bitflags = "= 1.3.2"
+fatfs = { path = "../rust-fatfs", version = "0.4", default-features = false, features = [
+    "alloc",
+    "lfn",
+] }
