+++ epoll/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "epoll"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+base_file = { path = "../base_file" }
+bitflags = "= 1.3.2"
+lock = { path = "../dependencies/kernel-sync" }
+numeric-enum-macro = { path = "../dependencies/numeric-enum-macro" }
