+++ epoll/src/epoll_file.rs
@@ -1,10 +1,10 @@
 //! epoll 类型文件
 
+use base_file::File;
 use lock::Mutex;
 use alloc::sync::Arc;
 use alloc::collections::{BTreeMap, BTreeSet};
-use super::{File, EpollEvent, EpollCtl};
-use crate::syscall::ErrorNo;
+use crate::{EpollEvent, EpollCtl, EpollErrorNo};
 
 /// 用作 epoll 的文件
 pub struct EpollFile {
@@ -37,13 +37,12 @@ impl EpollFile {
         }
     }
     /// 进行控制操作，如成功则返回 Ok(())，否则返回对应的错误编号
-    pub fn epoll_ctl(&self, op: EpollCtl, fd: i32, event: EpollEvent) -> Result<(), ErrorNo> {
-        info!("epool ctl: {:?}, fd: {}, event: {:?}", op, fd, event);
+    pub fn epoll_ctl(&self, op: EpollCtl, fd: i32, event: EpollEvent) -> Result<(), EpollErrorNo> {
         let list = &mut self.inner.lock().interest_list;
         match op {
             EpollCtl::ADD => {
                 if list.contains_key(&fd) {
-                    return Err(ErrorNo::EEXIST);
+                    return Err(EpollErrorNo::EEXIST);
                 } else {
                     list.insert(fd, event);
                 }
@@ -53,12 +52,12 @@ impl EpollFile {
                     // 根据 BTreeMap 的语义，这里的 insert 相当于把原来的值替换掉
                     list.insert(fd, event);
                 } else {
-                    return Err(ErrorNo::ENOENT);
+                    return Err(EpollErrorNo::ENOENT);
                 }
             },
             EpollCtl::DEL => {
                 if list.remove(&fd).is_none() {
-                    return Err(ErrorNo::ENOENT);
+                    return Err(EpollErrorNo::ENOENT);
                 }
             }
         }
@@ -74,9 +73,5 @@ impl File for EpollFile {
     /// epoll 文件不可直接写
     fn write(&self, _buf: &[u8]) -> Option<usize> {
         None
-    }
-    /// 如果这个文件对应的是一个 epoll，则获取 epoll 文件。否则，返回 None
-    fn get_epoll_fd(&self) -> Option<EpollFile> {
-        Some(self.clone())
     }
 }
