+++ kernel/Cargo.toml
@@ -24,3 +24,5 @@ fatfs = { path = "../rust-fatfs", version = "0.4", default-features = false, fea
 fscommon = { path = "../fscommon", version = "0.1.1" }
 sbi-rt = { path = "../dependencies/sbi-rt" }
 numeric-enum-macro = { path = "../dependencies/numeric-enum-macro" }
+epoll = { path = "../epoll" }
+base_file = { path = "../base_file" }
