--- kernel/src/file/epoll/mod.rs
@@ -1,9 +1,0 @@
-//! 和 epoll 相关的操作
-
-use super::File;
-
-
-mod flags;
-pub use flags::{EpollEvent, EpollEventType, EpollCtl};
-mod epoll_file;
-pub use epoll_file::EpollFile;
