+++ kernel/src/file/vfs/null.rs
@@ -1,7 +1,6 @@
 //! 空文件，用于 dev/null
 
-use super::{File, Kstat};
-use crate::file::{normal_file_mode, StMode};
+use base_file::{File, Kstat, normal_file_mode, StMode};
 
 pub struct NullFile;
