+++ kernel/src/file/vfs/zero.rs
@@ -1,7 +1,6 @@
 //! 另一种空文件，用于 dev/zero
 
-use super::{File, Kstat};
-use crate::file::{normal_file_mode, StMode};
+use base_file::{File, Kstat, normal_file_mode, StMode};
 
 pub struct ZeroFile;
