+++ kernel/src/syscall/mod.rs
@@ -21,11 +21,13 @@ mod socket;
 mod syscall_no;
 mod times;
 
+use base_file::Kstat;
 pub use flags::{PollFd, ErrorNo};
 use flags::*;
 use fs::*;
 use futex::*;
 pub use futex::{check_thread_blocked, wake_thread, set_waiter_for_thread};
+use epoll::EpollEvent;
 use loops::*;
 pub use loops::clear_loop_checker;
 use process::*;
@@ -34,7 +36,7 @@ use socket::*;
 use syscall_no::SyscallNo;
 use times::*;
 
-use crate::file::{FsStat, Kstat, EpollEvent};
+use crate::file::FsStat;
 use crate::signal::SigAction;
 use crate::task::ITimerVal;
 use crate::timer::{TimeSpec, TimeVal};
